+++ libs/ledger-live-common/src/hooks/recoverFeatureFlag.ts
@@ -52,10 +52,28 @@ export function useQuickAccessURI(
   return useReplacedURI(uri, id);
 }
 
+export function useAlreadyOnboardedURI(
+  servicesConfig: Feature_ProtectServicesMobile | null,
+): string | undefined {
+  const uri = servicesConfig?.params?.managerStatesData?.NEW?.alreadyOnboardedURI;
+  const id = servicesConfig?.params?.protectId;
+
+  return useReplacedURI(uri, id);
+}
+
 export function useAlreadySubscribedURI(
   servicesConfig: Feature_ProtectServicesMobile | null,
 ): string | undefined {
   const uri = servicesConfig?.params?.managerStatesData?.NEW?.alreadySubscribedURI;
+  const id = servicesConfig?.params?.protectId;
+
+  return useReplacedURI(uri, id);
+}
+
+export function useHomeURI(
+  servicesConfig: Feature_ProtectServicesMobile | null,
+): string | undefined {
+  const uri = servicesConfig?.params?.account?.homeURI;
   const id = servicesConfig?.params?.protectId;
 
   return useReplacedURI(uri, id);
@@ -79,7 +97,7 @@ export function useUpsellPath(
 }
 
 export function useLoginURI(
-  servicesConfig: Feature_ProtectServicesDesktop | null,
+  servicesConfig: Feature_ProtectServicesMobile | Feature_ProtectServicesDesktop | null,
 ): string | undefined {
   const uri = servicesConfig?.params?.account?.loginURI;
   const id = servicesConfig?.params?.protectId;
@@ -93,4 +111,78 @@ export function useLoginPath(
   const uri = useLoginURI(servicesConfig);
 
   return usePath(servicesConfig, uri);
+}
+
+export function useRestore24URI(
+  servicesConfig: Feature_ProtectServicesDesktop | null,
+): string | undefined {
+  const uri = servicesConfig?.params?.onboardingCompleted?.restore24URI;
+  const id = servicesConfig?.params?.protectId;
+
+  return useReplacedURI(uri, id);
+}
+
+export function useRestore24Path(
+  servicesConfig: Feature_ProtectServicesDesktop | null,
+): string | undefined {
+  const uri = useRestore24URI(servicesConfig);
+
+  return usePath(servicesConfig, uri);
+}
+
+export function useAccountURI(
+  servicesConfig: Feature_ProtectServicesDesktop | null,
+): string | undefined {
+  const uri = servicesConfig?.params?.account?.homeURI;
+  const id = servicesConfig?.params?.protectId;
+
+  return useReplacedURI(uri, id);
+}
+
+export function useAccountPath(
+  servicesConfig: Feature_ProtectServicesDesktop | null,
+): string | undefined {
+  const uri = useAccountURI(servicesConfig);
+
+  return usePath(servicesConfig, uri);
+}
+
+export function useAlreadySeededDeviceURI(
+  servicesConfig: Feature_ProtectServicesDesktop | null,
+): string | undefined {
+  const uri = servicesConfig?.params?.onboardingCompleted?.alreadyDeviceSeededURI;
+  const id = servicesConfig?.params?.protectId;
+
+  return useReplacedURI(uri, id);
+}
+
+export function useAlreadySeededDevicePath(
+  servicesConfig: Feature_ProtectServicesDesktop | null,
+): string | undefined {
+  const uri = useAlreadySeededDeviceURI(servicesConfig);
+
+  return usePath(servicesConfig, uri);
+}
+
+export function useCustomPath(
+  servicesConfig: Feature_ProtectServicesDesktop | Feature_ProtectServicesMobile | null,
+  page?: string,
+  source?: string,
+  deeplinkCampaign?: string,
+): string | undefined {
+  const modelUri = usePostOnboardingURI(servicesConfig);
+  const customUri = useMemo(() => {
+    const [basicUri] = modelUri ? modelUri.split("?") : [];
+    const uri = new URL(basicUri);
+
+    if (page) uri.searchParams.append("redirectTo", page);
+    if (source) uri.searchParams.append("source", source);
+    if (source && deeplinkCampaign) {
+      uri.searchParams.append("ajs_recover_source", source);
+      uri.searchParams.append("ajs_recover_campaign", deeplinkCampaign);
+    }
+    return uri;
+  }, [deeplinkCampaign, modelUri, page, source]);
+
+  return usePath(servicesConfig, customUri.toString()) ?? undefined;
 }